Halte Medienkomponenten innerhalb der Quality Gates
