+++ packages/studio-ui-react/src/studio-media-picker-overlay.tsx
@@ -1,7 +1,7 @@
-import { Button } from './button.js';
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from './dialog.js';
 import { MediaIntakePanel } from './media-intake-panel.js';
 import { StudioMediaPickerLibraryPanel } from './studio-media-picker-library-panel.js';
+import { StudioMediaPickerModeActions } from './studio-media-picker-mode-actions.js';
 import { StudioMediaPickerReviewPanel } from './studio-media-picker-review-panel.js';
 import type {
   StudioMediaPickerAssetDetail,
@@ -57,64 +57,6 @@ type StudioMediaPickerOverlayProps = Readonly<{
   feedbackTone?: 'default' | 'success' | 'error';
   labels: StudioMediaPickerOverlayLabels;
 }>;
-
-const StudioMediaPickerModeTabs = ({
-  labels,
-  mode,
-  onChangeMode,
-  onAddManual,
-  onClose,
-  disabled,
-}: Readonly<{
-  labels: StudioMediaPickerOverlayLabels['modes'];
-  mode: StudioMediaPickerMode;
-  onChangeMode: (mode: 'library' | 'upload') => void;
-  onAddManual: () => string | void;
-  onClose: () => void;
-  disabled: boolean;
-}>) => (
-  <div className="flex flex-wrap gap-2 border-b border-border/60 pb-4">
-    <Button
-      type="button"
-      disabled={disabled}
-      aria-pressed={mode === 'upload'}
-      onClick={() => onChangeMode('upload')}
-    >
-      {labels.upload}
-    </Button>
-    <Button
-      type="button"
-      disabled={disabled}
-      variant="secondary"
-      aria-pressed={mode === 'library'}
-      onClick={() => onChangeMode('library')}
-    >
-      {labels.library}
-    </Button>
-    <button
-      type="button"
-      disabled={disabled}
-      className="rounded-sm px-1 text-sm font-medium text-primary underline-offset-4 hover:underline focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50"
-      onClick={() => {
-        const uiId = onAddManual();
-        onClose();
-        if (uiId) {
-          globalThis.setTimeout(
-            () => globalThis.document?.getElementById(`content-media-${uiId}-url`)?.focus(),
-            0
-          );
-        }
-      }}
-    >
-      {labels.manual}
-    </button>
-    {mode === 'review' ? (
-      <Button type="button" variant="secondary" disabled>
-        {labels.review}
-      </Button>
-    ) : null}
-  </div>
-);
 
 const StudioMediaPickerOverlayBody = ({
   assets,
@@ -230,7 +172,7 @@ export const StudioMediaPickerOverlay = ({
           <DialogDescription>{labels.description}</DialogDescription>
         </DialogHeader>
 
-        <StudioMediaPickerModeTabs
+        <StudioMediaPickerModeActions
           disabled={isBusy}
           labels={labels.modes}
           mode={mode}
